backend

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 /* 
 Video.find({}, (error, videos) => {
@@ -12,7 +13,7 @@ export const home = async (req, res) => {
 
 export const watch = async (req, res) => {
   const { id } = req.params; // 데이터베이스에서 각 데이터의 id값 뽑아옴
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner");
   if (!video /* = video === null */) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
   }
@@ -21,7 +22,13 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
   const { id } = req.params; // 데이터베이스에서 각 데이터의 id값 뽑아옴
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   if (!video /* = video === null */) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
   }
@@ -29,11 +36,18 @@ export const getEdit = async (req, res) => {
 }; // form을 화면에 보여주는 녀석
 
 export const postEdit = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
+
   const { id } = req.params; // 데이터베이스에서 각 데이터의 id값 뽑아옴
   const { title, description, hashtags } = req.body;
   const video = await Video.exists({ _id: id });
   if (!video /* = video === null */) {
     return res.render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   await Video.findByIdAndUpdate(id, {
     title,
@@ -48,13 +62,22 @@ export const getUpload = (req, res) => {
 };
 
 export const postUpload = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
+  const { path: fileUrl } = req.file;
   const { title, description, hashtags } = req.body;
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
       description,
+      fileUrl,
+      owner: _id,
       hashtags: Video.formatHashtags(hashtags),
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect("/");
   } catch (error) {
     return res.status(400).render("upload", {
@@ -66,7 +89,20 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
+  const user = await User.findById(id);
+  if (!video /* = video === null */) {
+    return res.render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
+
   await Video.findByIdAndDelete(id);
+  user.videos.splice(user.videos.indexOf(id), 1);
   return res.redirect("/");
 };
